+++ util/spec_helper.rb
@@ -19,6 +19,7 @@ require_relative '../models/test_config'
 require_relative '../models/boac/boac_user'
 require_relative '../models/boac/appts_and_notes/timeline_record'
 require_relative '../models/boac/appts_and_notes/timeline_note_appt'
+require_relative '../models/boac/appts_and_notes/timeline_e_form'
 require_relative '../models/boac/advisor_role'
 require_relative '../models/boac/dept_membership'
 require_relative '../models/boac/academic_standing'
@@ -76,6 +77,7 @@ require_relative 'junction_utils'
 require_relative 'lrs_utils'
 require_relative 'nessie_utils'
 require_relative 'nessie_filter_utils'
+require_relative 'nessie_timeline_utils'
 require_relative 'oec_utils'
 require_relative 'squiggy_utils'
 require_relative 'suite_c_utils'
